requirements.txt
django
djangorestframework
pillow
gunicorn
python-dotenv
django-oauth-toolkit
django-cors-headers